requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0
matplotlib>=3.5.0
PyYAML>=6.0
ccxt>=3.0.0
scipy>=1.9.0
optuna>=3.0.0
shap>=0.41.0
websocket-client>=1.4.0
python-dotenv>=0.19.0
dill>=0.3.5
pandas-ta>=0.3.14b
EMD-signal>=1.6.0
psutil>=5.9.0
